requirements.txt
Flask>=3.1,<4.0
gunicorn>=21.2
torch
torchvision
numpy>=1.26,<2.0
Pillow
python-dotenv
